Removing shortcuts from Guard and using Against extensions throughout for consistency.
