+++ src/GuardClauses.UnitTests/GuardAgainstOutOfRange.cs
@@ -11,17 +11,7 @@ namespace GuardClauses.UnitTests
         [InlineData(1, 1, 3)]
         [InlineData(2, 1, 3)]
         [InlineData(3, 1, 3)]
-        public void DoesNothingGivenInRangeValueUsingShortcutMethod(int input, int rangeFrom, int rangeTo)
-        {
-            Guard.AgainsOutOfRange(input, "index", rangeFrom, rangeTo);
-        }
-
-        [Theory]
-        [InlineData(1, 1, 1)]
-        [InlineData(1, 1, 3)]
-        [InlineData(2, 1, 3)]
-        [InlineData(3, 1, 3)]
-        public void DoesNothingGivenInRangeValueUsingExtensionMethod(int input, int rangeFrom, int rangeTo)
+        public void DoesNothingGivenInRangeValue(int input, int rangeFrom, int rangeTo)
         {
             Guard.Against.OutOfRange(input, "index", rangeFrom, rangeTo);
         }
@@ -30,16 +20,7 @@ namespace GuardClauses.UnitTests
         [InlineData(-1, 1, 3)]
         [InlineData(0, 1, 3)]
         [InlineData(4, 1, 3)]
-        public void ThrowsGivenOutOfRangeValueUsingShortcutMethod(int input, int rangeFrom, int rangeTo)
-        {
-            Assert.Throws<ArgumentOutOfRangeException>(() => Guard.AgainsOutOfRange(input, "index", rangeFrom, rangeTo));
-        }
-
-        [Theory]
-        [InlineData(-1, 1, 3)]
-        [InlineData(0, 1, 3)]
-        [InlineData(4, 1, 3)]
-        public void ThrowsGivenOutOfRangeValueUsingExtensionMethod(int input, int rangeFrom, int rangeTo)
+        public void ThrowsGivenOutOfRangeValue(int input, int rangeFrom, int rangeTo)
         {
             Assert.Throws<ArgumentOutOfRangeException>(() => Guard.Against.OutOfRange(input, "index", rangeFrom, rangeTo));
         }
@@ -57,7 +38,7 @@ namespace GuardClauses.UnitTests
         [InlineData(-1, 3, 1)]
         [InlineData(0, 3, 1)]
         [InlineData(4, 3, 1)]
-        public void ThrowsGivenInvalidArgumentValueUsingExtensionMethod(int input, int rangeFrom, int rangeTo)
+        public void ThrowsGivenInvalidArgumentValue(int input, int rangeFrom, int rangeTo)
         {
             Assert.Throws<ArgumentException>(() => Guard.Against.OutOfRange(input, "index", rangeFrom, rangeTo));
         }
