+++ src/GuardClauses.UnitTests/NamespaceSeparate/ExtendingGuard.cs
@@ -3,6 +3,9 @@
 // By using the same namespace, the required using for Ardalis.GuardClauses will include custom guards regardless of location.
 namespace Ardalis.GuardClauses
 {
+    /// <summary>
+    /// An example Guard extension method. Throws if input is "foo".
+    /// </summary>
     public static class FooGuard
     {
         public static void Foo(this IGuardClause guardClause, string input, string parameterName)
