+++ src/GuardClauses/GuardClauseExtensions.cs
@@ -1,0 +1,70 @@
+﻿using System;
+
+namespace Ardalis.GuardClauses
+{
+    /// <summary>
+    /// A collection of common guard clauses, implented as extensions.
+    /// </summary>
+    /// <example>
+    /// Guard.Against.Null(input, nameof(input));
+    /// </example>
+    public static class GuardClauseExtensions
+    {
+        /// <summary>
+        /// Throws an ArgumentNullException if input is null.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <exception cref="ArgumentNullException"></exception>
+        public static void Null(this IGuardClause guardClause, object input, string parameterName)
+        {
+            if (null == input)
+            {
+                throw new ArgumentNullException(parameterName);
+            }
+        }
+
+        /// <summary>
+        /// Throws an ArgumentNullException if input is null.
+        /// Throws an ArgumentException if input is an empty string.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <exception cref="ArgumentNullException"></exception>
+        /// <exception cref="ArgumentException"></exception>
+        public static void NullOrEmpty(this IGuardClause guardClause, string input, string parameterName)
+        {
+            Guard.Against.Null(input, parameterName);
+            if (input == String.Empty)
+            {
+                throw new ArgumentException($"Required input {parameterName} was empty.", parameterName);
+            }
+        }
+
+        /// <summary>
+        /// Throws an ArgumentOutOfRange if input is less than <see cref="rangeFrom"/> or greater than <see cref="rangeTo"/>
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="rangeFrom"></param>
+        /// <param name="rangeTo"></param>
+        /// <exception cref="ArgumentException"></exception>
+        /// <exception cref="ArgumentOutOfRangeException"></exception>
+        public static void OutOfRange(this IGuardClause guardClause, int input, string parameterName,
+            int rangeFrom, int rangeTo)
+        {
+            if (rangeFrom > rangeTo)
+            {
+                throw new ArgumentException($"{nameof(rangeFrom)} should be less or equal than {nameof(rangeTo)}");
+            }
+
+            if (input < rangeFrom || input > rangeTo)
+            {
+                throw new ArgumentOutOfRangeException($"Input {parameterName} was out of range", parameterName);
+            }
+        }
+    }
+}
